+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 - Added cache writing to disk
-- Added `config` method
+- Added `setConfig` method
 
 ## [2.0.0] - 2017-08-13
 - Changed `set` method: last argument now is an object, see documentation.